requirements.txt
Django==1.11
Jinja2==2.10
MarkupSafe==1.0
coreapi==2.3.3
coreschema==0.0.4
djangorestframework==3.8.2
itypes==1.1.0
pytz==2018.5
uritemplate==3.0.0
whitenoise==3.3.1
